// src/app/models/utilisateur.ts
export class Utilisateur {
    id: number = 0;
    nom: string = "";
    prenom: string = "";
    email: string = "";
    password: string = "";
    adresse: string = "";
    telephone: string = "";
    roleID: number = 0;
    reseauID: number = 0;
    etat: string = "";
    createdAt: string = "";
}